Add more tests for RemoveSnapshots (#1223)
Previously none of the tests in TestRemoveSnapshots used the code path for
valid snapshots which contained references to manifests which were created
in a snapshot that had previously expired. Although there are no bugs in that
logic, we add two tests which check this pathway to protect against future
breakage.