Add a Parallelized Scan Planning Action
Planning for very large scans can take a considerable amount of time even for queries
that will not return a large amount of data. Some of the reason of this is that all
planning for Spark Reads happens locally prior to any data actually be fetched from Spark.
We can speed this up by providing a distributing planning Action which will distribute
the generation of the actual scan plan. This should allow us to scale planning with
compute and IO resources.