Core: Fix NullPointerException in ManifestReader (#1730)
This fixes a NullPointerException that is thrown by a ManifestReader for delete files when there is a query filter. The DeleteFileIndex projects all fields of a delete manifest, so it doesn't call select to select specific columns, unlike ManifestGroup, which selects * by default. When select is not called, methods that check whether to add stats columns fail, but only if there is a row filter because stats columns are not needed if there is no row filter.

Existing tests either called select to configure the reader, or didn't pass a row filter and projected all rows. This adds a test that uses DeleteFileIndex and a test for ManifestReader. This also fixes dropStats in addition to requireStatsProjection.

Co-authored-by: 钟保罗 <zhongbaoluo@shxgroup.net>

(cherry picked from commit d0625c8da5b22f6ef1e22059b3d6d847fbf9b350)