[Spark] Allows RewriteDataFiles to update partitioning (#337)
Previously the RewriteDataFiles code assumed that the data files were written with the correct partitioning for the the table. This causes an issue if the table's partition spec is evolved as some files will have old partition information. The contents of these files can not be assumed to end up in the same output partition and we need to do a shuffle to place it in the right files. To fix this we group all files which need to be repartitioned together and shuffle their contents so they end up in the correct output partitions.