Spark: Allow Bin-Pack optimization to work across different partition specs (#4279)
Previously, Bin-Pack would always use a NONE distribution for rewriting datafiles. This assumed
that incoming datafiles were from the same partition as the new output files would
be written to. When the spec is changed, the data must be resorted to place the data in the
new correct partitions.

(cherry picked from commit 610714e46741807356054a86efc653f5e3ddb167)