Internal (Boson): Make Boson `useDecimal128` and `useLazyMaterialization` configurable in Iceberg (#1061) (#1075)
This cherry-picks https://github.pie.apple.com/IPR/apache-incubator-iceberg/pull/1061 from `apple-1.3.x` to the `apple-1.3.0.x` branch.

We are seeing some customer issues that may related to lazy materialization. However, currently the config is hardcoded to `true` in Iceberg and there's no way to disable it. This PR enables us to do that.

Co-authored-by: Huaxin Gao <huaxin_gao@apple.com>