Flink: Ensure DynamicCommitter idempotence in the presence of failures (#1702)
Previously, the DynamicCommitter could commit duplicate WriteResults when
recovering from failures, leading to incorrect data in tables. This change gets
rid of the unnecessary commits for each WriteResult, which caused the problems
in the first place, and instead commits all the WriteResults for each Flink
checkpoint / table / branch together.

The WriteResults for each checkpoint / table / branch can safely be committed
together, even in the presence of delete files. Since we now commit once per
Flink checkpoint and table / branch, we can use the already existing max Flink
checkpoint id in the snapshot table properties for the table / branch to detect
whether we have already committed.

We intentionally chose not to combine WriteResults across Flink checkpoints,
which would be possible when a checkpoint only contains append-only data. While
technically feasible, it is a premature optimization which complicates the
implementation and maintainability for solving a very rare edge case of
committing multiple pending checkpoints. Multiple checkpoints can only occur
with a very low checkpoint interval or when concurrent checkpointing is
enabled. Even in such a situation, it is preferable to keep the commits
separately, as this makes it easier to reason about the applied changes.